extract Fs2StatefulServerCall to single file
